+++ package.json
@@ -18,11 +18,12 @@
     "browserify": "^17.0.0",
     "chart.js": "^2.9.4",
     "d3": "^6.7.0",
+    "jquery": "^3.6.0",
     "jws": "^4.0.0",
     "leaflet": "^1.7.1",
     "leaflet.markercluster": "^1.5.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-liquidchart": "^0.3.0",
     "rimraf": "^3.0.2",
     "watchify": "^4.0.0"
@@ -40,7 +41,7 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^14.14.35"
+    "@types/node": "^14.14.44"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
